セーブ

+++ PG3_05_01/scene/IScene.cpp
@@ -1,7 +1,7 @@
 ﻿#include "IScene.h"
 
 //タイトルシーンで初期化
-int IScene::sceneNo_ = static_cast<int>(SCENE::Title);
+int IScene::sceneNo_ = SCENE::Title;
 
 //仮想のデストラクタ
 IScene::~IScene(){
