+++ PG3_05_01/scene/gameSceneManager/game/StageScene.cpp
@@ -12,7 +12,7 @@ void StageScene::Initialize() {
 void StageScene::Update(char* keys, char* preKeys) {
 	gameObject_.position += gameObject_.velocity;
 	if (keys[DIK_SPACE] && !preKeys[DIK_SPACE]) {
-		sceneNo_ = static_cast<int>(End);
+		sceneNo_ = End;
 	}
 }
