+++ PG3_05_01/scene/gameSceneManager/title/TitleScene.cpp
@@ -13,7 +13,7 @@ void TitleScene::Initialize(){
 void TitleScene::Update(char* keys, char* preKeys){
 	gameObject_.position += gameObject_.velocity;
 	if (keys[DIK_SPACE] && !preKeys[DIK_SPACE]) {
-		sceneNo_ = static_cast<int>(Game_Stage);
+		sceneNo_ = Game_Stage;
 	}
 }
